Implement testing
After many hours in office hours, testing is finally finished. Most of my code is built in ofApp.cpp, and unfortunately due to Xcode bugs we cannot test any openFrameworks components. Since my code is built off of ofRectangles, which is an openFrameworks library, most of it is unable to be tested, but per office hours this should be enough, and since my game works properly we know most of the methods work regardless.